Update README.md with Enhanced Profile Information
This commit enhances the README.md file, providing a more comprehensive and visually appealing profile for Abdelfattah BOUHLALI. The update includes detailed social media links with badges, GitHub statistics, programming languages and technologies used, a dynamic gif, a snake animation, and a visitor badge. This aims to offer a better understanding of Abdelfattah's skills, interests, and contributions, creating a more engaging experience for visitors to the profile.